+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * NP-714 Upgrade Storybook to v6
+* NP-856 Fix focus styles in error summary
 
 ## <sub>v1.9.1</sub>